Rename the Upgrade class to Crowbar::UpgradeStatus
(cherry picked from commit 74b990e04796e668f9a320d7e850d9096d595f89)
